Moved all audio processing logic into audio stream
updated gitignore
