+++ src/audio/audio_stream.c
@@ -3,7 +3,15 @@
 
 LOG_MODULE_REGISTER(audio_stream);
 
-int pdm_config(AudioStream * audio_stream) {
+K_MSGQ_DEFINE(device_message_queue, sizeof(struct save_wave_msg), 8, 4);
+
+#define MEM_SLAB_BLOCK_COUNT 8
+
+K_MEM_SLAB_DEFINE_STATIC(mem_slab, MAX_BLOCK_SIZE, MEM_SLAB_BLOCK_COUNT, 4); //align mem slab to 4 bytes
+
+uint8_t writing = 0;
+
+int pdm_init(AudioStream * audio_stream) {
     if (!device_is_ready(audio_stream->dmic_ctx)) {
         LOG_ERR("%s is not ready", audio_stream->dmic_ctx);
         return 0;
@@ -11,7 +19,7 @@ int pdm_config(AudioStream * audio_stream) {
 
     struct pcm_stream_cfg stream = {
         .pcm_width = SAMPLE_BIT_WIDTH,
-        .mem_slab  = audio_stream->mem_slab,
+        .mem_slab  = &mem_slab,
     };
 
     struct dmic_cfg cfg = {
@@ -44,3 +52,82 @@ int pdm_config(AudioStream * audio_stream) {
     LOG_INF("LEFT GAIN: %d, RIGHT GAIN: %d", l_gain, r_gain);
     return 0;
 }
+
+void process_audio() {
+    struct save_wave_msg msg;
+
+    while(1) {
+        if (k_msgq_get(&device_message_queue, &msg, K_FOREVER) == 0) {
+            LOG_INF("Consumer: Received data %p\n", msg.buffer);
+            int ret = 0;
+            if(ret!=0) {
+                LOG_ERR("Failed to write wav header");
+            }
+            ret = write_wav_data(msg.audio_file, msg.buffer, msg.size);
+            //PROCESS AUDIO / DSP HERE:
+
+            k_mem_slab_free(&mem_slab, msg.buffer);
+            if (ret != 0) {
+                LOG_ERR("Failed to write to file, rc=%d", ret);
+                return;
+            }
+            LOG_INF("Written Data sucessfully");
+            writing = 0;
+
+        }
+    }
+}
+
+void stop_capture(struct device *dmic_dev, struct fs_file_t *audio_file) {
+    int ret = dmic_trigger(dmic_dev, DMIC_TRIGGER_STOP);
+    if (ret < 0) {
+        LOG_ERR("STOP trigger failed: %d", ret);
+    }
+    sd_card_close(audio_file);
+}
+
+//main thread
+int capture_audio(AudioStream *audio_stream)
+{
+int ret;
+struct save_wave_msg msg;
+
+ret = dmic_trigger(audio_stream->dmic_ctx, DMIC_TRIGGER_START);
+    if (ret < 0) {
+        LOG_ERR("START trigger failed: %d", ret);
+        return ret;
+    }
+
+for (int  i = 0; i < WAV_LENGTH_BLOCKS; i ++) {
+    ret = dmic_read(audio_stream->dmic_ctx, 0, &msg.buffer, &msg.size, READ_TIMEOUT);
+    if (ret < 0) {
+        LOG_ERR("%d - read failed: %d", i, ret);
+        return ret;
+    }
+    msg.audio_file = audio_stream->wav_config.wav_file;
+    LOG_INF("%d - got buffer %p of %u bytes", i, msg.buffer, msg.size);
+    writing = 1;
+    ret = k_msgq_put(&device_message_queue, &msg, K_NO_WAIT);
+
+    if (ret < 0) {
+    LOG_ERR("Failed to enqueue message: %d", ret);
+    return ret;
+    }
+}
+
+//@TO-DO, get rid of this, remove global writing flag
+while (1) {
+    if (writing == 0){
+        stop_capture(audio_stream->dmic_ctx, audio_stream->wav_config.wav_file);
+        break;
+    } else  {
+        k_sleep(K_MSEC(100));
+    }
+}
+
+LOG_INF("Audio Capture Finished");
+
+return ret;
+}
+
+K_THREAD_DEFINE(subscriber_task_id, CONFIG_MAIN_STACK_SIZE, process_audio, NULL, NULL, NULL, 3, 0, 0);
