+++ src/audio/audio_stream.h
@@ -3,6 +3,7 @@
 
 #include <zephyr/kernel.h>
 #include <zephyr/audio/dmic.h>
+#include "../modules/sd_card.h"
 #include "wav_file.h"
 #include <nrfx_pdm.h>
 #include "../macros.h"
@@ -10,10 +11,20 @@
 typedef struct {
     WavConfig wav_config;
     const struct device *const dmic_ctx;
-    struct k_mem_slab *mem_slab; 
+    //struct k_mem_slab *mem_slab; 
     nrf_pdm_gain_t pdm_gain;
 } AudioStream;
 
-int pdm_config(AudioStream * audio_stream);
+struct save_wave_msg {
+	void *buffer;
+	size_t size;
+	struct fs_file_t *audio_file
+};
+
+int pdm_init(AudioStream * audio_stream);
+
+int capture_audio(AudioStream *audio_stream);
+
+void process_audio();
 
 #endif
